[FLINK-17854][core] Move InputStatus to flink-core
That way, the InputStatus can be used by API classes that directly connect with the asynchronous
input paradigm of the streaming runtime.
